+++ src/launcher/features/apps/App/OpenApp.tsx
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import React from 'react';
+import React, { useState } from 'react';
 import Button from 'react-bootstrap/Button';
 import { logger } from '@nordicsemiconductor/pc-nrfconnect-shared';
 
@@ -28,17 +28,19 @@ import { type DisplayedApp, isInProgress } from '../appsSlice';
 
 const OpenApp: React.FC<{ app: DisplayedApp }> = ({ app }) => {
     const dispatch = useLauncherDispatch();
+    const [openning, setOpening] = useState<boolean>(false);
 
     if (!isInstalled(app)) return null;
 
     return (
         <Button
             title={`Open ${app.displayName}`}
-            disabled={isInProgress(app)}
+            disabled={openning || isInProgress(app)}
             onClick={async () => {
                 let updatedApp: InstalledDownloadableApp | undefined;
                 if (isDownloadable(app) && !isWithdrawn(app)) {
                     // The reason we don't include it in the compatibility check is due to needing the new app data for updated engine and jlink version in the compatibility check
+                    setOpening(true);
                     updatedApp = (await dispatch(checkForUpdatableApps([app]))
                         .then(async updatableApps => {
                             if (updatableApps.length === 1) {
@@ -62,6 +64,9 @@ const OpenApp: React.FC<{ app: DisplayedApp }> = ({ app }) => {
                         // Ignore if we fail to check for updates
                         .catch(err => {
                             logger.error(err);
+                        })
+                        .finally(() => {
+                            setOpening(false);
                         })) as InstalledDownloadableApp | undefined; // Typescript is unable to type this correctly
                 }